feat(issues): Lazy load trace waterfall components
We have both the old and new trace views directly bundled into issue details. Attempt to keep these split in their own chunks to avoid dragging in the entire app on page load.
